requirements.txt
aiohttp==3.8.1
Django==4.0.6
djangorestframework==3.13.1
django-jazzmin==2.5.0
gunicorn==20.1.0
whitenoise==6.2.0
